cir-bpm-mlra/backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
passlib[bcrypt]
python-jose
tensorflow
numpy
pydantic